pyproject.toml
[project]
name = "openai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.4",
    "ipykernel>=6.29.5",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]
